mode MDImage to its own file
